cambio

+++ app/controllers/homes_controller.rb
@@ -40,12 +40,13 @@ class HomesController < ApplicationController
 
 
   def destroy
-	@home=Home.find(params[:id])
-  	if @home.destroy
-		redirect_to homes_path
-	  else
-		redirect_to homes_path
- 	  end
-  end
-
-end
+    id = params[:id]
+    @home=Home.find(id)
+    if @home.reservation
+       if @home.destroy
+		     redirect_to homes_path
+	      else
+		      redirect_to homes_path
+ 	     end
+    end
+   end
